+++ examples/dagster_examples/intro_tutorial/custom_types.py
@@ -4,7 +4,7 @@ from dagster import DagsterType, execute_pipeline, pipeline, solid
 
 SimpleDataFrame = DagsterType(
     name='SimpleDataFrame',
-    type_check_fn=lambda value: isinstance(value, list),
+    type_check_fn=lambda _, value: isinstance(value, list),
     description='A naive representation of a data frame, e.g., as returned by csv.DictReader.',
 )