+++ examples/dagster_examples/intro_tutorial/custom_types_4.py
@@ -14,7 +14,7 @@ from dagster import (
 )
 
 
-def less_simple_data_frame_type_check(value):
+def less_simple_data_frame_type_check(_, value):
     if not isinstance(value, list):
         return TypeCheck(
             success=False,