+++ python_modules/dagster/dagster/core/types/dagster_type.py
@@ -19,11 +19,13 @@ class DagsterType(object):
     '''Define a type in dagster. These can be used in the inputs and outputs of solids.
 
     Args:
-        type_check_fn (Callable[[Any], [Union[bool, TypeCheck]]]):
+        type_check_fn (Callable[[TypeCheckContext, Any], [Union[bool, TypeCheck]]]):
             The function that defines the type check. It takes the value flowing
             through the input or output of the solid. If it passes, return either
             `True` or a `TypeCheck` with `success` set to `True`. If it fails,
             return either `False` or a `TypeCheck` with `success` set to `False`.
+            The first argument must be named context (or _, _context, context_).
+            Use `required_resource_keys` for access to resources.
         key (Optional[str]): The unique key to identify types programatically.
             The key property always has a value. If you omit key to the argument
             to the init function, it instead receives the value of `name`. If
@@ -59,6 +61,7 @@ class DagsterType(object):
             argument. In these cases the serialization_strategy argument is not sufficient because
             serialization requires specialized API calls, e.g. to call an S3 API directly instead
             of using a generic file object. See ``dagster_pyspark.DataFrame`` for an example.
+        required_resource_keys (Optional[Set[str]]): Resource keys required by the type_check_fn.
         is_builtin (bool): Defaults to False. This is used by tools to display or
             filter built-in types (such as String, Int) to visually distinguish
             them from user-defined types. Meant for internal use.
@@ -75,6 +78,7 @@ class DagsterType(object):
         output_materialization_config=None,
         serialization_strategy=None,
         auto_plugins=None,
+        required_resource_keys=None,
     ):
         check.opt_str_param(key, 'key')
         check.opt_str_param(name, 'name')
@@ -110,8 +114,12 @@ class DagsterType(object):
             SerializationStrategy,
             PickleSerializationStrategy(),
         )
+        self.required_resource_keys = check.opt_set_param(
+            required_resource_keys, 'required_resource_keys',
+        )
 
         self._type_check_fn = check.callable_param(type_check_fn, 'type_check_fn')
+        _validate_type_check_fn(self._type_check_fn, self.name)
 
         auto_plugins = check.opt_list_param(auto_plugins, 'auto_plugins', of_type=type)
 
@@ -130,8 +138,8 @@ class DagsterType(object):
             'All types must have a valid display name, got None for key {}'.format(key),
         )
 
-    def type_check(self, value):
-        retval = self._type_check_fn(value)
+    def type_check(self, context, value):
+        retval = self._type_check_fn(context, value)
 
         if not isinstance(retval, (bool, TypeCheck)):
             raise DagsterInvariantViolationError(
@@ -188,6 +196,35 @@ class DagsterType(object):
         return False
 
 
+def _validate_type_check_fn(fn, name):
+    from dagster.seven import get_args
+
+    args = get_args(fn)
+
+    # py2 doesn't filter out self
+    if len(args) >= 1 and args[0] == 'self':
+        args = args[1:]
+
+    if len(args) == 2:
+        possible_names = {
+            '_',
+            'context',
+            '_context',
+            'context_',
+        }
+        if args[0] not in possible_names:
+            DagsterInvalidDefinitionError(
+                'type_check function on type "{name}" must have first '
+                'argument named "context" (or _, _context, context_).'.format(name=name,)
+            )
+        return True
+
+    raise DagsterInvalidDefinitionError(
+        'type_check function on type "{name}" must take 2 arguments, '
+        'received {count}.'.format(name=name, count=len(args))
+    )
+
+
 class BuiltinScalarDagsterType(DagsterType):
     def __init__(self, name, type_check_fn, *args, **kwargs):
         super(BuiltinScalarDagsterType, self).__init__(
@@ -198,7 +235,7 @@ class BuiltinScalarDagsterType(DagsterType):
     def is_scalar(self):
         return True
 
-    def type_check_method(self, _value):
+    def type_check_method(self, _context, _value):
         raise NotImplementedError()
 
 
@@ -211,7 +248,7 @@ class _Int(BuiltinScalarDagsterType):
             type_check_fn=self.type_check_method,
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         return _fail_if_not_of_type(value, six.integer_types, 'int')
 
 
@@ -238,7 +275,7 @@ class _String(BuiltinScalarDagsterType):
             type_check_fn=self.type_check_method,
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         return _fail_if_not_of_type(value, six.string_types, 'string')
 
 
@@ -251,7 +288,7 @@ class _Path(BuiltinScalarDagsterType):
             type_check_fn=self.type_check_method,
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         return _fail_if_not_of_type(value, six.string_types, 'string')
 
 
@@ -264,7 +301,7 @@ class _Float(BuiltinScalarDagsterType):
             type_check_fn=self.type_check_method,
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         return _fail_if_not_of_type(value, float, 'float')
 
 
@@ -277,7 +314,7 @@ class _Bool(BuiltinScalarDagsterType):
             type_check_fn=self.type_check_method,
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         return _fail_if_not_of_type(value, bool, 'bool')
 
 
@@ -305,7 +342,7 @@ class Anyish(DagsterType):
             auto_plugins=auto_plugins,
         )
 
-    def type_check_method(self, _value):
+    def type_check_method(self, _context, _value):
         return TypeCheck(success=True)
 
     @property
@@ -358,7 +395,7 @@ class _Nothing(DagsterType):
     def is_nothing(self):
         return True
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         if value is not None:
             return TypeCheck(
                 success=False,
@@ -409,7 +446,7 @@ class PythonObjectDagsterType(DagsterType):
             key=key, name=name, type_check_fn=self.type_check_method, **kwargs
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         if not isinstance(value, self.python_type):
             return TypeCheck(
                 success=False,
@@ -475,8 +512,10 @@ class OptionalType(DagsterType):
     def display_name(self):
         return self.inner_type.display_name + '?'
 
-    def type_check_method(self, value):
-        return TypeCheck(success=True) if value is None else self.inner_type.type_check(value)
+    def type_check_method(self, context, value):
+        return (
+            TypeCheck(success=True) if value is None else self.inner_type.type_check(context, value)
+        )
 
     @property
     def is_nullable(self):
@@ -528,13 +567,13 @@ class ListType(DagsterType):
     def display_name(self):
         return '[' + self.inner_type.display_name + ']'
 
-    def type_check_method(self, value):
+    def type_check_method(self, context, value):
         value_check = _fail_if_not_of_type(value, list, 'list')
         if not value_check.success:
             return value_check
 
         for item in value:
-            item_check = self.inner_type.type_check(item)
+            item_check = self.inner_type.type_check(context, item)
             if not item_check.success:
                 return item_check
 
@@ -586,7 +625,7 @@ class Stringish(DagsterType):
     def is_scalar(self):
         return True
 
-    def type_check_method(self, value):
+    def type_check_method(self, _context, value):
         return _fail_if_not_of_type(value, six.string_types, 'string')