+++ python_modules/dagster/dagster/core/types/python_set.py
@@ -44,7 +44,7 @@ class _TypedPythonSet(DagsterType):
             type_check_fn=self.type_check_method,
         )
 
-    def type_check_method(self, value):
+    def type_check_method(self, context, value):
         from dagster.core.definitions.events import TypeCheck
 
         if not isinstance(value, set):
@@ -56,7 +56,7 @@ class _TypedPythonSet(DagsterType):
             )
 
         for item in value:
-            item_check = self.item_type.type_check(item)
+            item_check = self.item_type.type_check(context, item)
             if not item_check.success:
                 return item_check