+++ python_modules/dagster/dagster_tests/core_tests/runtime_types_tests/test_python_tuple.py
@@ -83,18 +83,18 @@ def test_open_typing_tuple_input_fail():
 def test_typed_python_tuple_directly():
     int_str_tuple = create_typed_tuple(int, str)
 
-    int_str_tuple.type_check((1, 'foo'))
+    int_str_tuple.type_check(None, (1, 'foo'))
 
-    res = int_str_tuple.type_check(None)
+    res = int_str_tuple.type_check(None, None)
     assert not res.success
 
-    res = int_str_tuple.type_check('bar')
+    res = int_str_tuple.type_check(None, 'bar')
     assert not res.success
 
-    res = int_str_tuple.type_check((1, 2, 3))
+    res = int_str_tuple.type_check(None, (1, 2, 3))
     assert not res.success
 
-    res = int_str_tuple.type_check(('1', 2))
+    res = int_str_tuple.type_check(None, ('1', 2))
     assert not res.success
 
 
@@ -103,15 +103,15 @@ def test_nested_python_tuple_directly():
 
     nested_tuple = create_typed_tuple(bool, list, int_str_tuple_kls)
 
-    nested_tuple.type_check((True, [1], (1, 'foo')))
+    nested_tuple.type_check(None, (True, [1], (1, 'foo')))
 
-    res = nested_tuple.type_check(None)
+    res = nested_tuple.type_check(None, None)
     assert not res.success
 
-    res = nested_tuple.type_check('bar')
+    res = nested_tuple.type_check(None, 'bar')
     assert not res.success
 
-    res = nested_tuple.type_check((True, [1], (1, 2)))
+    res = nested_tuple.type_check(None, (True, [1], (1, 2)))
     assert not res.success